src/resources/query_lambda/requirements.txt
psycopg2-binary==2.9.4